refactor: move SuggestionBuilder
